feat(remote): fractional per-file progress for fetch/pull/push
The transfer bar advanced 0 -> 1 per file, so a single large blob sat at
0/1 the whole time. Drive the bar fractionally instead: advance by
bytes_in_chunk / file_size as each file streams, so a half-done file
shows ~0.5/N and the percentage climbs continuously (tqdm can even show
an ETA now). Per-file size is free: stat for uploads, ContentLength for
downloads.

A shared _FractionalProgress accumulator (completed_files + sum of
in-flight fractions) keeps the value exact at the end. The byte/rate
postfix is retained as the textual detail, and the displayed count is
rounded so it reads "0.50/1" rather than a raw float.

Verified against local minio: a 2GB single-file fetch fills 4% -> 96% ->
100% smoothly.

+++ packages/pivot/src/pivot/remote/sync.py
@@ -427,7 +427,7 @@ async def _push_async(
     remote_name: str,
     targets: list[str] | None = None,
     jobs: int | None = None,
-    callback: Callable[[int, int, str], None] | None = None,
+    callback: Callable[[float, int, str], None] | None = None,
     all_stages: dict[str, RegistryStageInfo] | None = None,
     byte_callback: Callable[[int], None] | None = None,
 ) -> TransferSummary:
@@ -470,7 +470,7 @@ async def _push_async(
     if callback is not None:
         name_by_hash = build_hash_path_index(state_dir, all_stages, project.get_project_root())
 
-        def _translate(completed: int, total: int, ident: str) -> None:
+        def _translate(completed: float, total: int, ident: str) -> None:
             callback(completed, total, name_by_hash.get(ident, ident[:8]))
 
         upload_callback = _translate
@@ -502,7 +502,7 @@ def push(
     remote_name: str,
     targets: list[str] | None = None,
     jobs: int | None = None,
-    callback: Callable[[int, int, str], None] | None = None,
+    callback: Callable[[float, int, str], None] | None = None,
     all_stages: dict[str, RegistryStageInfo] | None = None,
     byte_callback: Callable[[int], None] | None = None,
 ) -> TransferSummary:
@@ -531,7 +531,7 @@ async def _pull_async(
     remote_name: str,
     targets: list[str] | None = None,
     jobs: int | None = None,
-    callback: Callable[[int, int, str], None] | None = None,
+    callback: Callable[[float, int, str], None] | None = None,
     all_stages: dict[str, RegistryStageInfo] | None = None,
     exclude_patterns: list[str] | None = None,
     byte_callback: Callable[[int], None] | None = None,
@@ -567,7 +567,7 @@ async def _pull_async(
     if callback is not None:
         name_by_hash = build_hash_path_index(state_dir, all_stages, project.get_project_root())
 
-        def _translate(completed: int, total: int, ident: str) -> None:
+        def _translate(completed: float, total: int, ident: str) -> None:
             callback(completed, total, name_by_hash.get(ident, ident[:8]))
 
         download_callback = _translate
@@ -603,7 +603,7 @@ def pull(
     remote_name: str,
     targets: list[str] | None = None,
     jobs: int | None = None,
-    callback: Callable[[int, int, str], None] | None = None,
+    callback: Callable[[float, int, str], None] | None = None,
     all_stages: dict[str, RegistryStageInfo] | None = None,
     exclude_patterns: list[str] | None = None,
     byte_callback: Callable[[int], None] | None = None,
